+++ lib/src/blocs/page/close_event.dart
@@ -4,4 +4,4 @@ import 'event.dart';
 ///
 /// The page could be anywhere in the stack as long as at least one other page
 /// is left in the stack afterwards.
-class ScreenBlocCloseEvent extends ScreenBlocEvent {}
+class PageBlocCloseEvent extends PageBlocEvent {}