+++ lib/src/blocs/page/configuration_changed_event.dart
@@ -1,0 +1,6 @@
+import 'event.dart';
+
+/// Emit this when configuration of the screen changes.
+///
+/// The configuration will then be read with [PageBloc.currentConfiguration].
+class PageBlocConfigurationChangedEvent extends PageBlocEvent {}