+++ lib/src/blocs/page/event.dart
@@ -1,5 +1,5 @@
-/// The base class for [ScreenBloc] events.
+/// The base class for [PageBloc] events.
 ///
 /// Extend this for your custom events like emitting notifications
 /// to be shown in overlays etc.
-abstract class ScreenBlocEvent {}
+abstract class PageBlocEvent {}