+++ lib/src/blocs/page_stack/page_stack.dart
@@ -5,23 +5,28 @@ import 'package:flutter/widgets.dart';
 import 'package:rxdart/rxdart.dart';
 
 import 'event.dart';
-import 'screen_bloc_normalized_state.dart';
-import 'screen_event.dart';
-import 'page_stack_bloc_normalized_state.dart';
-import '../screen/close_event.dart';
-import '../screen/configuration_changed_event.dart';
-import '../screen/event.dart';
+import 'page_event.dart';
+import 'configuration.dart';
+import '../page/configuration.dart';
+import '../page/close_event.dart';
+import '../page/configuration_changed_event.dart';
+import '../page/event.dart';
 import '../../pages/abstract.dart';
 
 /// The source of pages for [Navigator] widget.
 ///
 /// [C] is the base class for all app's page configurations.
-class PageStackBloc<C> {
+class PageStackBloc<C extends PageConfiguration> {
   final _pages = <AbstractPage<C>>[];
   List<AbstractPage<C>> get pages => _pages;
 
   /// A factory to create pages from their serialized states.
+  /// It is called when a popped page should be re-created on back-front
+  /// navigation. If null, pages will not be re-created this way.
+  /// This makes navigation useless except for popping.
   final AbstractPage<C>? Function(String key, Map<String, dynamic> state)? createPage;
+
+  /// What to do if pushing a page with a key already existing in the stack.
   final DuplicatePageKeyAction duplicatePageKeyAction;
 
   final _eventsController = BehaviorSubject<PageStackBlocEvent>();
@@ -35,10 +40,10 @@ class PageStackBloc<C> {
     _pushNoFire(bottomPage);
   }
 
-  /// The first non-null page currentConfiguration from top.
-  C get topPageCurrentConfiguration {
+  /// The first non-null page configuration from top.
+  C getTopPageConfiguration() {
     for (final page in _pages.reversed) {
-      final configuration = page.currentConfiguration;
+      final configuration = page.getConfiguration();
       if (configuration != null) return configuration;
     }
 
@@ -52,7 +57,13 @@ class PageStackBloc<C> {
   }
 
   void _pushNoFire(AbstractPage<C> page) {
-    final oldPage = _findSameKeyPage(page.key);
+    final key = page.key;
+    if (key == null) {
+      _pushNewPageNoFire(page);
+      return;
+    }
+
+    final oldPage = _findSameKeyPage(key);
     if (oldPage == null) {
       _pushNewPageNoFire(page);
       return;
@@ -61,7 +72,7 @@ class PageStackBloc<C> {
     _pushDuplicateNoFire(oldPage, page);
   }
 
-  AbstractPage<C>? _findSameKeyPage(LocalKey key) {
+  AbstractPage<C>? _findSameKeyPage(ValueKey<String> key) {
     return _pages.firstWhereOrNull((page) => page.key == key);
   }
 
@@ -74,21 +85,21 @@ class PageStackBloc<C> {
     }
   }
 
-  void _onPageEvent(AbstractPage<C> page, ScreenBlocEvent event) {
+  void _onPageEvent(AbstractPage<C> page, PageBlocEvent event) {
     _emitPageEvent(page, event);
 
-    if (event is ScreenBlocCloseEvent && _pages.length >= 2) {
+    if (event is PageBlocCloseEvent && _pages.length >= 2) {
       _pages.remove(page);
       _firePageConfigurationChange(_pages.last);
       _schedulePageDisposal(page);
     }
   }
 
-  void _emitPageEvent(AbstractPage<C> page, ScreenBlocEvent event) {
-    final pageStackEvent = PageStackScreenBlocEvent<C>(
+  void _emitPageEvent(AbstractPage<C> page, PageBlocEvent event) {
+    final pageStackEvent = PageStackPageBlocEvent<C>(
       page: page,
       bloc: page.bloc,
-      screenBlocEvent: event,
+      pageBlocEvent: event,
     );
     _eventsController.sink.add(pageStackEvent);
   }
@@ -110,16 +121,16 @@ class PageStackBloc<C> {
         _pages.add(oldPage);
         break;
     }
+
+    throw Exception('Unknown duplicatePageKeyAction: $duplicatePageKeyAction');
   }
 
   void _firePageConfigurationChange(AbstractPage<C> page) {
     _eventsController.sink.add(
-      PageStackScreenBlocEvent(
+      PageStackPageBlocEvent(
         page: page,
         bloc: page.bloc,
-        screenBlocEvent: ScreenBlocConfigurationChangedEvent(
-          //configuration: page.currentConfiguration,
-        ),
+        pageBlocEvent: PageBlocConfigurationChangedEvent(),
       ),
     );
   }
@@ -157,24 +168,20 @@ class PageStackBloc<C> {
     return false;
   }
 
-  /// Normalizes all pages' states serialization.
-  PageStackBlocNormalizedState get normalizedState {
-    return PageStackBlocNormalizedState(
-      screenStates: _pages.map((p) => _getPageNormalizedState(p)).toList(growable: false),
-    );
-  }
-
-  ScreenBlocNormalizedState _getPageNormalizedState(AbstractPage<C> page) {
-    return ScreenBlocNormalizedState(
-      pageKey: page.key.toString(),
-      state: page.bloc?.normalizedState ?? const <String, dynamic>{},
+  /// Surveys all pages' configurations for serialization.
+  PageStackConfiguration getConfiguration() {
+    return PageStackConfiguration(
+      pageConfigurations: _pages
+          .map((p) => p.getConfiguration())
+          .toList(growable: false),
     );
   }
 
   /// Recovers pages and their states.
   ///
   /// 1. Compares the current stack with given page states from bottom to top.
-  ///    For each page with key matching state's pageKey, recovers its state.
+  ///    For each page with key matching configuration's key, recovers its
+  ///    state.
   ///
   /// 2. When a mismatch is found, pops and disposes the remaining pages.
   ///    This goes from top to bottom.
@@ -182,16 +189,25 @@ class PageStackBloc<C> {
   /// 3. For page states that were not matched, creates the pages
   ///    with [createPage] factory and sets their states.
   ///    Stops at the first page that failed to be created.
-  set normalizedState(PageStackBlocNormalizedState state) {
+  void setConfiguration(PageStackConfiguration configuration, {bool fire = true}) {
     int matchedIndex = 0;
-    int matchLength = min(_pages.length, state.screenStates.length);
+    int matchLength = min(_pages.length, configuration.pageConfigurations.length);
 
     for (; matchedIndex < matchLength; matchedIndex++) {
       final page = _pages[matchedIndex];
-      final screenState = state.screenStates[matchedIndex];
+      final pc = configuration.pageConfigurations[matchedIndex];
 
-      if (page.key.value != screenState.pageKey) break;
-      page.bloc?.normalizedState = screenState.state;
+      if (page.key == null && pc == null) {
+        // A page without key is implied to be the same, and no state can
+        // be applied to it.
+        continue;
+      }
+
+      if (page.key?.value != pc?.key) break;
+
+      if (pc != null) {
+        page.bloc?.setStateMap(pc.state);
+      }
     }
 
     for (int i = _pages.length; --i >= matchedIndex; ) {
@@ -199,27 +215,50 @@ class PageStackBloc<C> {
       _schedulePageDisposal(page);
     }
 
-    for (int i = matchedIndex; i < state.screenStates.length; i++) {
-      if (!_createPage(state.screenStates[i])) break;
+    for (int i = matchedIndex; i < configuration.pageConfigurations.length; i++) {
+      final pc = configuration.pageConfigurations[i];
+      if (pc == null) {
+        // Pages without configuration are transient dialogs, these are OK
+        // to skip. Otherwise we will never be able to recover good URLed pages
+        // on top of transient dialogs.
+        continue;
+      }
+
+      if (!_createPage(pc)) {
+        // But if we cannot create a page with configuration, it is not OK.
+        // Configuration implies recoverability.
+        // Also consider throwing Exception in createPage factory on failure
+        // so we will not get here.
+        break;
+      }
     }
 
-    _firePageConfigurationChange(_pages.last);
+    if (fire) _firePageConfigurationChange(_pages.last);
+
+    // TODO: Prevent emptying. Maybe keep the list of popped pages
+    //       and only dispose them when we determine the stack is not emptied.
+    if (_pages.isEmpty) {
+      throw Exception(
+        'PageStackBloc is emptied by setting a configuration state. '
+        'The stack should never be empty according to the Navigator API.'
+      );
+    }
   }
 
-  bool _createPage(ScreenBlocNormalizedState state) {
-    if (createPage == null) return false;
+  bool _createPage(PageConfiguration pc) {
+    if (createPage == null || pc.factoryKey == null) return false;
 
-    final page = createPage!(state.pageKey, state.state);
+    final page = createPage!(pc.factoryKey!, pc.state);
     if (page == null) return false;
 
-    page.bloc?.normalizedState = state.state;
+    page.bloc?.setStateMap(pc.state);
     _pushNoFire(page);
     return true;
   }
 
   void _schedulePageDisposal(AbstractPage<C> page) async {
     // If we dispose the page immediately (or even with 1 frame delay of
-    // Future.delayed(Duration.zero), the bloc will dispose
+    // Future.delayed(Duration.zero)), the bloc will dispose
     // TextEditingController objects, and we get the exception of using disposed
     // controllers before the screen is gone.
     // TODO: Find a guaranteed synchronous way.