+++ lib/src/pages/material.dart
@@ -1,26 +1,36 @@
 import 'package:flutter/material.dart';
 
 import 'abstract.dart';
-import '../blocs/screen/screen.dart';
+import '../blocs/page/configuration.dart';
+import '../blocs/page/page.dart';
 
-abstract class AbstractMaterialPage<C> extends MaterialPage implements AbstractPage<C> {
-  final ValueKey _nonNullKey;
-
-  @override
-  ValueKey get key => _nonNullKey;
+abstract class AbstractMaterialPage<C extends PageConfiguration> extends MaterialPage implements AbstractPage<C> {
+  final ValueKey<String>? _valueKey;
 
   @override
-  ScreenBloc<C, dynamic>? get bloc => null;
+  ValueKey<String>? get key => _valueKey;
+
+  final String? _factoryKey;
+
+  // Don't use getter because 'factoryKey' is a convenient identifier
+  // of a constant for user subclasses.
+  @override
+  String? getFactoryKey() => _factoryKey ?? _valueKey?.value;
 
   @override
-  C? get currentConfiguration => null;
+  PageBloc<C>? get bloc => null;
+
+  @override
+  C? getConfiguration() => null;
 
   const AbstractMaterialPage({
-    required ValueKey key,
+    ValueKey<String>? key,
+    String? factoryKey,
     required Widget child,
   })
       :
-        _nonNullKey = key,
+        _valueKey = key,
+        _factoryKey = factoryKey,
         super(
           key: key,
           child: child,