+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.1.2
+
+* **BREAKING**: Renamed `AppBlocNormalizedState` to `AppConfiguration`.
+* **BREAKING**: Renamed `PageStackBlocNormalizedState` to `PageStackConfiguration`.
+* **BREAKING**: Renamed `ScreenBlocNormalizedState` to `PageConfiguration`.
+* **BREAKING**: Renamed `Screen...` events to `Page...` events.
+* **BREAKING**: Extracted `PageBloc.states` to the new `PageStatefulBloc` subclass.
+* Allow `null` for page keys.
+* Added docs.
+
 ## 0.1.1
 
 * **BREAKING**: Removed `ScreenBlocConfigurationChangedEvent.configuration`.